feat: rename `prodserver` command to `server` (deprecate `prodserver`)
Introduces a new `server` management command that takes over as the primary
way to start configured production processes. The original `prodserver`
command is retained as a deprecated subclass that emits a DeprecationWarning
on every invocation, giving users a clear migration window.

The `PRODUCTION_PROCESSES` setting and the `django_prodserver` package name
are intentionally unchanged. The `prodserver` alias will be removed in
django-prodserver 4.0.0.

Bumps version to 3.0.0 and updates README, docs, and autodoc accordingly.
