+++ pyproject.toml
@@ -4,7 +4,7 @@ requires = ["setuptools"]
 
 [project]
 name = "django-prodserver"
-version = "2.4.0"
+version = "3.0.0"
 description = "A management command to start production servers/workers with a consistent interface"
 readme = "README.md"
 license = { text = "MIT" }
