+++ src/django_prodserver/__init__.py
@@ -1,1 +1,1 @@
-__version__ = "2.4.0"
+__version__ = "3.0.0"
